fixed react unique key error & splitted Card from CardComponent
SavePagination
CardComponent
SWOTAnalysisMatrix
HoverWindow: fixed "hover" trigger error; added "focus" aswell
NormalClassifying
ClassifyingCardList
EvaluationComponent
UtilResultComponent
CreateDescriptionModal
