+++ src/components/tools/swot-analysis/steps/SWOTClassifyAlternativeActions/NormalClassifying/ClassifyingCardList/ClassifyingCardList.tsx
@@ -10,20 +10,27 @@ interface ClassifyingCardListProps {
 
 class ClassifyingCardList extends PureComponent<ClassifyingCardListProps, {}> {
     render() {
-        return <>
-            <div className={"actionCards"}>
-                {this.props.actions.map((action) => {
-                    return <ClassifyingCard action={action}
-                                            onChangeClick={this.props.onOpenClassificationModalClick}/>;
-                })}
-            </div>
+        return (
+            <>
+                <div className={"actionCards"}>
+                    {this.props.actions.map((action, index) => {
+                        return (
+                            <ClassifyingCard
+                                key={action.name + "-" + action.index + "-" + index}
+                                action={action}
+                                onChangeClick={this.props.onOpenClassificationModalClick}
+                            />
+                        );
+                    })}
+                </div>
 
-            {(this.props.actions.length <= 0) && (
-                <span>
-                    Keine Handlungsalternativen zugeordnet...
-                </span>
-            )}
-        </>;
+                {(this.props.actions.length <= 0) && (
+                    <span>
+                        Keine Handlungsalternativen zugeordnet...
+                    </span>
+                )}
+            </>
+        );
     }
 }
