+++ src/general-components/HoverWindow/HoverWindow.tsx
@@ -41,12 +41,11 @@ class HoverWindow extends PureComponent<HoverWindowProps, HoverWindowState> {
 
 
     static defaultProps = {
-        trigger: "hover",
+        trigger: ["hover", "focus"],
         disabled: false
     }
 
     render() {
-
         const {customPopUp, children, trigger, onToggle, placement, ...customProps} = this.props;
 
         // early exit
@@ -71,6 +70,7 @@ class HoverWindow extends PureComponent<HoverWindowProps, HoverWindowState> {
                     </Popover.Body>) : undefined}
             </Popover>
         );
+
         return (
             <OverlayTrigger trigger={this.props.trigger} overlay={popupFunction ?? popover}
                             onToggle={this.props.onToggle} placement={placement}>
