+++ src/general-components/Tool/Home/SaveResourceList/SavePagination/SavePagination.tsx
@@ -50,7 +50,7 @@ class SavePagination extends Component<SaveResourceListProps, SavePaginationStat
 
                     {page?.data.map(save => {
                         return (
-                            <SaveCard save={save} toolLink={this.props.tool!.getLink()} onTrash={() => {
+                            <SaveCard key={save.tool_id + " " +  save.id} save={save} toolLink={this.props.tool!.getLink()} onTrash={() => {
                                 this.props.savesControlCallbacks.deleteSave(save);
                             }}/>
                         );
